Move IPC code for appDetails to shared
